Revert "Merge branch 'main' of https://github.com/quevedoana/Spa-Relax-G14"
This reverts commit 7786e49a0faadfb113a4207e8a32ed714ee761a4, reversing
changes made to 476b747acf69b4eaca030902ed1282473e822ce4.

+++ src/Persistencia/ClienteData.java
@@ -64,7 +64,7 @@ public class ClienteData {
             if (resultado.next()) {
 
                 cli = new Cliente(
-                        resultado.getLong("DNI"), 
+                        resultado.getInt("DNI"), 
                         resultado.getString("NombreCompleto"), 
                         resultado.getLong("Telefono"),
                         resultado.getInt("Edad"), 
@@ -95,7 +95,7 @@ public class ClienteData {
             ResultSet resultado = ps.executeQuery();
             if (resultado.next()) {
 
-                cli = new Cliente(resultado.getLong("DNI"), resultado.getString("NombreCompleto"), resultado.getLong("Telefono"), resultado.getInt("Edad"), resultado.getString("Afecciones"), resultado.getBoolean("Estado"));
+                cli = new Cliente(resultado.getLong("DNI"), resultado.getString("NombreCompleto"), resultado.getInt("Telefono"), resultado.getInt("Edad"), resultado.getString("Afecciones"), resultado.getBoolean("Estado"));
                 cli.setCodCli(resultado.getInt("CodCli"));
 
             } else {
